host/requirements.txt
python-telegram-bot>=21.0
croniter>=3.0.0
pyyaml>=6.0
aiohttp>=3.9.0
aiofiles>=23.0

# Slack
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# Discord
discord.py>=2.3.0

# Gmail / Google APIs
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
